+++ NEWS.rst
@@ -6,6 +6,7 @@ of every change, see the Git log.
 
 Latest
 ------
+* Minor: Added shallow Full RLNC Coders.
 * Major: Upgrade to kodo 23
 * Major: Upgrade to fifi 18
 * Major: Upgrade to sak 14